Update 2022-03-17
First expl3 version

The code is mainly copied from that of "minimalist", and there are still a lot of works to to. Notably, the configuration for the styles of theorem-like environments remains to be done.

The current style is named "fancy". When the (planned) cleaner design is available, this shall no longer be the default style.